requirements.txt
Flask
Flask-Cors
Flask-Bcrypt
python-dotenv
psycopg2-binary
PyJWT


